pyproject.toml
[project]
name = "add-product-to-order"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.1",
    "alembic-utils>=0.8.8",
    "asyncio>=4.0.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "loguru>=0.7.3",
    "orjson>=3.11.6",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.46",
    "sqlalchemy-utils>=0.42.1",
    "uvicorn>=0.40.0",
]
